Removed unnecessary Triangle Simplified 2D-clipping

+++ RasterizerConfig.h
@@ -27,8 +27,8 @@ namespace softlit
 
 	struct Viewport
 	{
-		float x;
-		float y;
+		uint32_t x;
+		uint32_t y;
 
 		uint32_t width;
 		uint32_t height;
@@ -39,6 +39,6 @@ namespace softlit
 		CullMode cullMode = CullMode::CULL_DISABLED;
 		VertexWinding vertexWinding = VertexWinding::CLOCKWISE;
 
-		Viewport viewport = { 0.f, 0.f, 1024u, 768u };
+		Viewport viewport = { 0u, 0u, 1024u, 768u };
 	};
 }
